+++ app/controllers/topics/posts_controller.rb
@@ -1,8 +1,9 @@
 class Topics::PostsController < ApplicationController
-
+before_filter :authenticate_user!
+  
   def show
     @topic = Topic.find(params[:topic_id])
-    authorize! :read, @topic, message: "You need to be signed in to do that."
+    authorize @topic
     @post = Post.find(params[:id])
     @comments = @post.comments
     @comment = Comment.new
@@ -11,13 +12,14 @@ class Topics::PostsController < ApplicationController
   def new
     @topic = Topic.find(params[:topic_id])
     @post = Post.new
-    authorize! :create, Post, message: "You need to be a member to create a new post."
+    authorize @post
   end
 
   def create
     @topic = Topic.find(params[:topic_id])
-    @post = current_user.posts.build(params[:post])
-    authorize! :create, @post, message: "You need to be signed up to do that."
+    @post = current_user.posts.build(params.require(:post).permit(:title, :body))
+    @post.topic = @topic 
+    authorize @post
     if @post.save
     flash[:notice] = "Post was saved."
     redirect_to [@topic, @post]
@@ -30,16 +32,16 @@ class Topics::PostsController < ApplicationController
   def edit
     @topic = Topic.find(params[:topic_id])
     @post = Post.find(params[:id])
-    authorize! :edit, @post, message: "You need to own the post to edit it."
+    authorize @post
   end
 
   def update
-    @post = Post.find(params[:id])
     @topic = Topic.find(params[:topic_id])
-    authorize! :update, @post, message: "You need to own the post to edit it."
-    if @post.update_attributes(params[:post])
+    @post = Post.find(params[:id])
+    authorize @post
+    if @post.update_attributes(params.require(:post).permit(:title, :body))
       flash[:notice] = "Post was updated."
-      redirect_to @post
+      redirect_to [@topic, @post]
     else
       flash[:error] = "There was an error saving the post. Please try again."
       render :edit
@@ -49,9 +51,12 @@ class Topics::PostsController < ApplicationController
   def destroy
     @topic = Topic.find(params[:topic_id])
     @post = Post.find(params[:id])
+    authorize @post
+    @comments = @post.comments
+    @comment = Comment.new
 
     title = @post.title
-    authorize! :destroy, @post, message: "You need to own the post to delete it."
+    authorize @post
     if @post.destroy
       flash[:notice] = "\"#{title}\" was deleted successfully."
       redirect_to @topic
