+++ app/controllers/topics_controller.rb
@@ -1,27 +1,28 @@
 class TopicsController < ApplicationController
   def index
     @topics = Topic.visible_to(current_user).paginate(page: params[:page], per_page: 10)
+    authorize @topics
   end
 
   def new
     @topic = Topic.new
-    authorize! :create, @topic, message: "You need to be an admin to do that."
+    authorize @topic
   end
 
   def show
     @topic = Topic.find(params[:id])
-    authorize! :read, @topic, message: "You need to be signed in to do that."
+    authorize @topic
     @posts = @topic.posts.includes(:user).includes(:comments).paginate(page: params[:page], per_page: 10)
   end
 
   def edit
     @topic = Topic.find(params[:id])
-    authorize! :update, @topic, message: "You need to be an admin to do that."
+    authorize @topic
   end
 
   def create
-    @topic = Topic.new(params[:topic])
-    authorize! :create, @topic, message: "You need to be an admin to do that."
+    @topic = Topic.new(params.require(:topic).permit(:name, :description, :public))
+    authorize @topic
     if @topic.save
       redirect_to @topic, notice: "Topic was saved successfully."
     else
@@ -32,8 +33,8 @@ class TopicsController < ApplicationController
 
   def update
     @topic = Topic.find(params[:id])
-    authorize! :update, @topic, message: "You need to own the topic to update it."
-    if @topic.update_attributes(params[:topic])
+    authorize @topic
+    if @topic.update_attributes(topic_params)
       redirect_to @topic, notice: "Topic was saved successfully."
     else
       flash[:error] = "There was an error saving the topic. Please try again."
@@ -44,7 +45,7 @@ class TopicsController < ApplicationController
   def destroy
     @topic = Topic.find(params[:id])
     name = @topic.name
-    authorize! :destroy, @topic, message: "You need to own the topic to delete it."
+    authorize @topic
     if @topic.destroy
       flash[:notice] = "\"#{name}\" was deleted successfully."
       redirect_to topics_path
@@ -52,5 +53,12 @@ class TopicsController < ApplicationController
       flash[:error] = "There was an error deleting the topic."
       render :show
     end
+  end
+  def topic_params
+    params.require(:topic).permit(
+      :name,
+      :description,
+      :public
+    )
   end
 end
